refactor: chat을 talk으로 변경
