+++ android/app/src/main/res/drawable/ic_talk.xml
