+++ android/app/src/main/res/layout/fragment_talk.xml
@@ -5,7 +5,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.chat.ChatFragment">
+        tools:context=".presentation.talk.TalkFragment">
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
