+++ android/app/src/main/res/menu/menu_bottom_navigation.xml
@@ -9,9 +9,9 @@
         android:icon="@drawable/ic_stats"
         android:title="@string/bottom_navigation_stats" />
     <item
-        android:id="@+id/item_chat"
-        android:icon="@drawable/ic_chat"
-        android:title="@string/bottom_navigation_chat" />
+        android:id="@+id/item_talk"
+        android:icon="@drawable/ic_talk"
+        android:title="@string/bottom_navigation_talk" />
     <item
         android:id="@+id/item_challenge"
         android:icon="@drawable/ic_challenge"
